requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.6
cryptography==40.0.1
cssselect==1.2.0
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
distlib==0.3.6
Django==4.2.2
django-ckeditor==6.6.1
django-js-asset==2.0.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.89.0
ffmpy==0.3.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
gradio==3.16.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.20.1
ipython==8.8.0
itemadapter==0.8.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
lxml==4.9.2
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mediapipe==0.10.1
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openai==0.26.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.1
orjson==3.8.4
outcome==1.2.0
packaging==22.0
pandas==1.5.2
parsel==1.7.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.8.0
prompt-toolkit==3.0.36
Protego==0.2.1
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
PyDispatcher==2.0.7
pydub==0.25.1
pygame==2.3.0
Pygments==2.14.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
Scrapy==2.8.0
selenium==4.8.3
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.22.0
tldextract==3.4.0
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.10.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.4.0
tzdata==2023.3
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.23.1
w3lib==2.1.1
wcwidth==0.2.5
webdriver-manager==3.8.6
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
zope.interface==6.0
